requirements.txt
google-cloud-storage>=1.30.0
numpy>=1.19.0
pandas>=1.1
pymc3==3.11.0
pytz
scipy
xarray
xlrd==1.2.0
